Implement immediate-taking `Zbt0p93` instructions
These instructions take a `shamt`-sized immediate value in-place of rs2.

In binutils, it is common to allow the non-immediate version of an
opcode to take an immediate value and become the immediate-taking
version in the emitted code, e.g `sll a0,a0,12` will become the `slli`
opcode. This is a convenience for assembly programmers. This is
implemented using aliases as is done here.

The relative order of the opcodes in this list matters too, it should be
in the order:

- immediate-taking instruction
- non-immediate-taking instruction
- non-immediate-taking instruction, taking immediate, marked as an alias

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>